Increase window size to 1280x720 Upgrade all art to proper pixel art. Almost.

+++ source/SpaceMarine/SpaceMarineGame.cs
@@ -6,8 +6,9 @@ namespace DeenGames.SpaceMarine
     public class SpaceMarineGame : PuffinGame
     {
         public SpaceMarineGame() :
-            base(Constants.MAP_TILES_WIDE * Constants.TILE_WIDTH * Constants.GAME_ZOOM,
-                (Constants.MAP_TILES_HIGH * Constants.TILE_HEIGHT * Constants.GAME_ZOOM) + Constants.STATUS_BAR_HEIGHT)
+            base(
+                (int)(Constants.MAP_TILES_WIDE * Constants.TILE_WIDTH * Constants.GAME_ZOOM),
+                (int)(Constants.MAP_TILES_HIGH * Constants.TILE_HEIGHT * Constants.GAME_ZOOM) + Constants.STATUS_BAR_HEIGHT)
         {
         }
